More generic per-domain settings backend.
I was having a look at adding a cookie whitelist and it looked to be a
pretty small change if only the domain settings module where a little
more generic. So I copied off of the existing settings framework.

Generic storing of `[url][option] = value` with some special handling
for when value is a bool.

There is a couple of ways this could be integrated into the existing
settings framework with some special case-ing.

PS: For anyone already using this patch (@paretje ?) you need to change
`script toggle` to `domain javascript!` in keys.conf and
`enable-javascript` to `javascript` in domains.yaml.