Move domain specific JS and zoom policies to AbstractTab.
So they can be applied to either backend. Uses the mapping in websettings to
change properties.

The domain specific referer policy setting wasn't added to the webengine
backend because as far as I can tell setting the referer header isn't
supported by qtwebengine. Referer is stripped when using info.setHttpHeader in
the request interceptor in URLRequestHttpJob::Start() in webengine so that
plugins don't override privacy mode. set". But I can't find in any settings or
policy class. It looks like it can be set in the request object in webengine
and when opening a new url in a view and as a per-frame policy but that isn't
exposed in QT. You can use --qt-flag no-referrers to disable for the whole
session though.