WebEngine webview emit new signal on navigation accept.
Previously sitting the local javascript policy on a frame was done on
urlChanged like with webkit. But with webengine that is fired too late
to have an effect on the ongoing page load. So as an alternative a new
signal, new_navigation is added and the js policy is set when that
fires.

Also undo moving the js policy application up to the abstract tab for
obvious reasons.